ACS TMCDB with MySQL...change in "baciidl" module, "BACI.xsd" file
BACI.xsd file defines the default values for some generic properties (like Graph_max, Graph_min, Max_step etc.). In TMCDB with MySQL configuration, these default values are read from this BACI.xsd file and put into the created MySQL tables. Some of the default values caused problems while placing them into MySQL tables, throwing an SQL exception, indicating that "+/- infinity is not defined in MySQL". So some of the default values of the following properties were changed as follows;

** Graph_max = from 1.7976931348623157E+308  to  3.4028234663852886E+38
** Graph_min = from -1.7976931348623157E+308  to  -3.4028234663852886E+38
** Max_value = from 1.7976931348623157E+308  to  3.4028234663852886E+38
** Min_value = from -1.7976931348623157E+308  to  -3.4028234663852886E+38

The problem cause by the definition of maxiumum/minimum value of the double properties. Double max and min values are defined as +/-1.7976931348623157E+308, where MySQL accepts the values only between +/-3.4028234663852886E+38, which is the max/min value for the 'float'. If the given value is out of this interval, MySQL accepts the value as " +/- Infinity ", and it is not defined in MySQL. Changing these values solved the problem.