ACS TMCDB with MySQL...change in "cdb_rdb" module, "Ejb3TypeDeclaration.ftl" class
-- Editing the *ftl file for creating the table classes (Putting @Table annotation name and @Table annotation column names into backquote <TableName> --> `<TableName>` ).

** Hibernate server is responsible for putting the CDB configuration located within the .xml files into MySQL tables. For this operation (putting or getting the information from MySQL tables), new java classes are generated under the name of the corresponding MySQL table names (ex: Manager.java, Component.java, Schemas.java etc. under the directory cdb_rdb/src/gen/alma/acs/tmcdb) . Hibernate server identifies the table and column names by using the java annotations located within this generated classes.  

Example
@Table(name="SCHEMAS", uniqueConstraints = @UniqueConstraint(columnNames={"URN", "CONFIGURATIONID"}))

In that point, TMCDB faces with a problem originated by MySQL itself. Because MySQL has some predefined reserved table and column names for internal use, which coincidences with the table and column names that are required and have to be created for using TMCDB. One example for the reserved table names in MySQL is "Schemas" table that is shown above. MySQL only allows to use that kind of reserved table names within backquote character (`<TableName>`). Otherwise an exception will be thrown by MySQL Driver.

So @Table(name="SCHEMAS") will cause a problem and has to be transformed to @Table(name="`SCHEMAS`"), this transformation is also requried for the reserved column names. By editing this *.ftl file which is used for creating the MySQL table java classes, one can overcome this problem by putting all the table and column names within backquote, which is totally acceptable for MySQL.  

In this edited *ftl file "Ejb3TypeDecleration.ftl", all the generated table names and column name within @Table annotation was put into backquotes by hardcoding.